requirements.txt
# LLM & Agent Framework
langchain==0.1.0
langchain-openai==0.0.2
langgraph==0.0.20
langchain-core==0.1.10

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Utilities
loguru==0.7.2
python-dotenv==1.0.0
pyyaml==6.0.1
httpx==0.25.2
requests==2.31.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Code Analysis & Security
ast-comments==1.1.2
astroid==3.0.2

# Docker SDK (for sandbox)
docker==7.0.0

# Progress & CLI
tqdm==4.66.1
click==8.1.7
rich==13.7.0

